15-rag-evaluation/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# RAG Evaluation
ragas>=0.1.0  # RAG Assessment framework
langsmith>=0.1.0  # LangSmith for evaluation (optional but recommended)

# For environment variable management
python-dotenv>=1.0.0
